Add migration to create products database table
This is an auto generated migration. We need to add some more columns to
the table, and perhaps opt to use UUID IDs instead of auto increment
values that tie the Product domain model to the database. Although the
latter can hardly be completely avoided when using Laravel and Eloquent.