Implement Order entity as AggregateRoot
Making Order an aggregate root, we persist it by saving all business
events that have happened for it, and retrieve it again by replaying all
these business events.

Let's also provide a `findOrCreate()` method that allows us to set the
OrderId property on the instance, because the aggregate UUID is a
private property on AggregateRoot.