Return standard JSON validation error object
The validator can throw a validation exception which is caught in the
exception handler and transformed to a standard validation error
response. Let's use that one instead of building our own format.

In order to test this, we do need to make sure to pass `Accept:
application/json` as a request header.