Hydrate LineItem instances with actual Product instance
We're again not using an injected repository here but use the
convenience of the Product model's query builder. Eventually, to
separate concerns, we can refactor this to use a repository later.

A (hidden) business rule (or assumption if you will), is that Products
are not to be deleted when no longer in use or being sold. So we can't
run into a ProductDoesNotExist exception here. If we want to mark
Products as obsolete, we should make them soft deletable.