Create Order without using aggregate root fake
Using the `fake()` helper seems to be more of a hindrance than a help.
We can just create a new Order instance and manually record our events
onto it. This gives way easier code in our test and it does not require
the aggregate to actually be persisted. After all we're not testing
events here, we're only testing an Order to be serialized into an array.