Allow Money value objects to be zero
The total amount of value in an Order can be zero. Let's make this
possible in the VO.

This will also make it possible to have a Product with a price of 0, so
either we need a more specific type for prices, or simply allow a price
of 0 for free products.